+++ test/Npgsql.Tests/Types/DateTimeTests.cs
@@ -8,7 +8,7 @@ using static Npgsql.Tests.TestUtil;
 namespace Npgsql.Tests.Types
 {
     // Since this test suite manipulates TimeZone, it is incompatible with multiplexing
-    public class DateTimeTests : TypeTestBase
+    public class DateTimeTests : TestBase
     {
         #region Date
