+++ test/Npgsql.Tests/Types/GeometricTypeTests.cs
@@ -10,7 +10,7 @@ namespace Npgsql.Tests.Types
     /// <remarks>
     /// https://www.postgresql.org/docs/current/static/datatype-geometric.html
     /// </remarks>
-    class GeometricTypeTests : MultiplexingTypeTestBase
+    class GeometricTypeTests : MultiplexingTestBase
     {
         [Test]
         public Task Point()
