+++ test/Npgsql.Tests/Types/LegacyDateTimeTests.cs
@@ -10,7 +10,7 @@ namespace Npgsql.Tests.Types
 {
     // Since this test suite manipulates TimeZone, it is incompatible with multiplexing
     [NonParallelizable]
-    public class LegacyDateTimeTests : TypeTestBase
+    public class LegacyDateTimeTests : TestBase
     {
         [Test]
         public Task Timestamp_with_all_DateTime_kinds([Values] DateTimeKind kind)
